`examples` changed to `examples2`
