--- examples/src/main/scala/sttp/tapir/examples/testing/AkkaServerStubInterpreterExample.scala
@@ -1,67 +1,0 @@
-package sttp.tapir.examples.testing
-
-import org.scalatest.flatspec.AsyncFlatSpec
-import org.scalatest.matchers.should.Matchers
-import sttp.client3._
-import sttp.client3.testing.SttpBackendStub
-import sttp.model.StatusCode
-import sttp.tapir._
-import sttp.tapir.server.ServerEndpoint
-import sttp.tapir.server.akkahttp.AkkaHttpServerOptions
-import sttp.tapir.server.interceptor.exception.ExceptionHandler
-import sttp.tapir.server.interceptor.CustomiseInterceptors
-import sttp.tapir.server.model.ValuedEndpointOutput
-import sttp.tapir.server.stub.TapirStubInterpreter
-
-import scala.concurrent.{ExecutionContext, Future}
-
-class AkkaServerStubInterpreterExample extends AsyncFlatSpec with Matchers {
-
-  it should "use custom exception handler" in {
-    val stubBackend: SttpBackend[Future, Any] = TapirStubInterpreter(UsersApi.options, SttpBackendStub.asynchronousFuture)
-      .whenServerEndpoint(UsersApi.greetUser)
-      .thenThrowException(new RuntimeException("error"))
-      .backend()
-
-    sttp.client3.basicRequest
-      .get(uri"http://test.com/api/users/greet")
-      .send(stubBackend)
-      .map(_.body shouldBe Left("failed due to error"))
-  }
-
-  it should "run greet users logic" in {
-    val stubBackend: SttpBackend[Future, Any] = TapirStubInterpreter(UsersApi.options, SttpBackendStub.asynchronousFuture)
-      .whenServerEndpoint(UsersApi.greetUser)
-      .thenRunLogic()
-      .backend()
-
-    val response = sttp.client3.basicRequest
-      .get(uri"http://test.com/api/users/greet")
-      .header("Authorization", "Bearer password")
-      .send(stubBackend)
-
-    // then
-    response.map(_.body shouldBe Right("hello user123"))
-  }
-}
-
-object UsersApi {
-
-  val greetUser: ServerEndpoint[Any, Future] = endpoint.get
-    .in("api" / "users" / "greet")
-    .securityIn(auth.bearer[String]())
-    .out(stringBody)
-    .errorOut(stringBody)
-    .serverSecurityLogic(token =>
-      Future.successful {
-        if (token == "password") Right("user123") else Left("unauthorized")
-      }
-    )
-    .serverLogic(user => _ => Future.successful(Right(s"hello $user")))
-
-  val exceptionHandler = ExceptionHandler.pure[Future](ctx =>
-    Option(ValuedEndpointOutput(stringBody.and(statusCode), (s"failed due to ${ctx.e.getMessage}", StatusCode.InternalServerError)))
-  )
-  def options(implicit ec: ExecutionContext): CustomiseInterceptors[Future, AkkaHttpServerOptions] =
-    AkkaHttpServerOptions.customiseInterceptors.exceptionHandler(exceptionHandler)
-}
