+++ examples2/src/main/resources/logback.xml
