+++ examples2/src/main/resources/webapp/index.html
