+++ examples2/src/main/scala/sttp/tapir/examples2/HelloWorldAkkaServer.scala
@@ -1,0 +1,37 @@
+package sttp.tapir.examples2
+
+import akka.actor.ActorSystem
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.server.Route
+import sttp.tapir._
+import sttp.tapir.server.akkahttp.AkkaHttpServerInterpreter
+
+import scala.concurrent.{Await, Future}
+import scala.concurrent.duration._
+import sttp.client3._
+
+object HelloWorldAkkaServer extends App {
+  implicit val actorSystem: ActorSystem = ActorSystem()
+  import actorSystem.dispatcher
+
+  // the endpoint: single fixed path input ("hello"), single query parameter
+  // corresponds to: GET /hello?name=...
+  val helloWorld: PublicEndpoint[String, Unit, String, Any] =
+    endpoint.get.in("hello").in(query[String]("name")).out(stringBody)
+
+  // converting an endpoint to a route (providing server-side logic); extension method comes from imported packages
+  val helloWorldRoute: Route =
+    AkkaHttpServerInterpreter().toRoute(helloWorld.serverLogicSuccess(name => Future.successful(s"Hello, $name!")))
+
+  // starting the server
+  val bindAndCheck = Http().newServerAt("localhost", 8080).bindFlow(helloWorldRoute).map { _ =>
+    // testing
+    val backend: SttpBackend[Identity, Any] = HttpURLConnectionBackend()
+    val result: String = basicRequest.response(asStringAlways).get(uri"http://localhost:8080/hello?name=Frodo").send(backend).body
+    println("Got result: " + result)
+
+    assert(result == "Hello, Frodo!")
+  }
+
+  Await.result(bindAndCheck.transformWith { r => actorSystem.terminate().transform(_ => r) }, 1.minute)
+}
